+++ modules/test/arquillian-bridge/src/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,0 +1,1 @@
+com.liferay.arquillian.bridge.extension.LiferayArquillianExtension
