--- modules/test/liferay-arquillian-extension/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Arquillian Extension
-Bundle-SymbolicName: com.liferay.arquillian.extension
-Bundle-Version: 1.0.0
-Include-Resource: classes
