--- modules/test/liferay-arquillian-extension/src/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,0 @@
-com.liferay.arquillian.extension.LiferayArquillianExtension
