--- modules/test/liferay-arquillian-extension/src/com/liferay/arquillian/deployment/JUnitBridgeAuxiliaryArchiveAppender.java
@@ -1,46 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.arquillian.deployment;
-
-import com.liferay.arquillian.extension.LiferayArquillianExtension;
-import com.liferay.arquillian.junit.observer.JUnitBridgeObserver;
-
-import org.jboss.arquillian.container.test.spi.RemoteLoadableExtension;
-import org.jboss.arquillian.container.test.spi.client.deployment.AuxiliaryArchiveAppender;
-import org.jboss.shrinkwrap.api.Archive;
-import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.asset.EmptyAsset;
-import org.jboss.shrinkwrap.api.spec.JavaArchive;
-
-/**
- * @author Shuyang Zhou
- */
-public class JUnitBridgeAuxiliaryArchiveAppender
-	implements AuxiliaryArchiveAppender {
-
-	@Override
-	public Archive<?> createAuxiliaryArchive() {
-		JavaArchive archive = ShrinkWrap.create(
-			JavaArchive.class, "arquillian-junit-bridge.jar");
-
-		archive.add(EmptyAsset.INSTANCE, "/arquillian.remote.marker");
-		archive.addAsServiceProviderAndClasses(
-			RemoteLoadableExtension.class, LiferayArquillianExtension.class);
-		archive.addClass(JUnitBridgeObserver.class);
-
-		return archive;
-	}
-
-}
